requirements.txt
Flask==3.1.2
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
python-json-logger==4.0.0
psycopg[binary]==3.3.2
python-dotenv==1.2.1
requests==2.32.5
PyJWT==2.10.1
cryptography==46.0.3
pytest==7.4.0
pytest-cov==4.1.0